move form field code and StandaloneField
