+++ src/frontend/lib/components/forms/StandaloneField.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useMemo } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 
-import type { ApiFormFieldType } from '@lib/types/Forms';
+import type { ApiFormFieldType } from '../../types/Forms';
 import { ApiFormField } from './fields/ApiFormField';
 
 export function StandaloneField({
