+++ src/frontend/lib/components/forms/fields/AutoFillRightSection.tsx
