+++ src/frontend/lib/components/forms/fields/ChoiceField.tsx
@@ -1,8 +1,8 @@
-import type { ApiFormFieldType } from '@lib/types/Forms';
 import { Select } from '@mantine/core';
 import { useId } from '@mantine/hooks';
 import { useCallback, useMemo, useRef } from 'react';
 import type { FieldValues, UseControllerReturn } from 'react-hook-form';
+import type { ApiFormFieldType } from '../../../types/Forms';
 
 /**
  * Render a 'select' field for selecting from a list of choices
