+++ src/frontend/lib/components/forms/fields/DateField.tsx
@@ -1,10 +1,10 @@
-import type { ApiFormFieldType } from '@lib/types/Forms';
 import { t } from '@lingui/core/macro';
 import { DateInput } from '@mantine/dates';
 import dayjs from 'dayjs';
 import customParseFormat from 'dayjs/plugin/customParseFormat';
 import { useCallback, useId, useMemo } from 'react';
 import type { FieldValues, UseControllerReturn } from 'react-hook-form';
+import type { ApiFormFieldType } from '../../../types/Forms';
 
 dayjs.extend(customParseFormat);
