+++ src/frontend/lib/components/forms/fields/TableField.tsx
@@ -5,10 +5,10 @@ import { IconExclamationCircle } from '@tabler/icons-react';
 import { type ReactNode, useCallback, useEffect, useMemo } from 'react';
 import type { FieldValues, UseControllerReturn } from 'react-hook-form';
 
-import { AddItemButton } from '@lib/components/AddItemButton';
-import { identifierString } from '@lib/functions/Conversion';
-import type { ApiFormFieldType } from '@lib/types/Forms';
-import { InvenTreeIcon } from '../../../functions/icons';
+import { InvenTreeIcon } from '../../../../src/functions/icons';
+import { AddItemButton } from '../../../components/AddItemButton';
+import { identifierString } from '../../../functions/Conversion';
+import type { ApiFormFieldType } from '../../../types/Forms';
 import { StandaloneField } from '../StandaloneField';
 
 export interface TableFieldRowProps {
