+++ src/frontend/src/components/editors/TemplateEditor/TemplateEditor.tsx
@@ -26,13 +26,13 @@ import type React from 'react';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
 import { Boundary } from '@lib/components/Boundary';
+import { StandaloneField } from '@lib/components/forms/StandaloneField';
 import { ModelInformationDict } from '@lib/enums/ModelInformation';
 import { ModelType } from '@lib/enums/ModelType';
 import { apiUrl } from '@lib/functions/Api';
 import { api } from '../../../App';
 import type { TemplateI } from '../../../tables/settings/TemplateTable';
 import { SplitButton } from '../../buttons/SplitButton';
-import { StandaloneField } from '../../forms/StandaloneField';
 
 type EditorProps = {
   template: TemplateI;
