+++ src/frontend/src/components/wizards/OrderPartsWizard.tsx
@@ -1,6 +1,7 @@
 import { ActionButton } from '@lib/components/ActionButton';
 import { AddItemButton } from '@lib/components/AddItemButton';
 import { CopyButton } from '@lib/components/CopyButton';
+import { StandaloneField } from '@lib/components/forms/StandaloneField';
 import { ApiEndpoints } from '@lib/enums/ApiEndpoints';
 import { ModelType } from '@lib/enums/ModelType';
 import { apiUrl } from '@lib/functions/Api';
@@ -34,7 +35,6 @@ import { usePurchaseOrderFields } from '../../forms/PurchaseOrderForms';
 import useWizard from '../../hooks/UseWizard';
 import { RenderPartColumn } from '../../tables/ColumnRenderers';
 import RemoveRowButton from '../buttons/RemoveRowButton';
-import { StandaloneField } from '../forms/StandaloneField';
 import Expand from '../items/Expand';
 
 /**
