+++ src/frontend/src/forms/ReturnOrderForms.tsx
@@ -8,18 +8,18 @@ import {
 } from '@tabler/icons-react';
 import { useMemo } from 'react';
 
+import { StandaloneField } from '@lib/components/forms/StandaloneField';
 import { ApiEndpoints } from '@lib/enums/ApiEndpoints';
 import { ModelType } from '@lib/enums/ModelType';
 import RemoveRowButton from '../components/buttons/RemoveRowButton';
-import { StandaloneField } from '../components/forms/StandaloneField';
 
+import type { TableFieldRowProps } from '@lib/components/forms/fields/TableField';
 import { apiUrl } from '@lib/functions/Api';
 import { useCreateApiFormModal } from '@lib/hooks/UseForm';
 import type {
   ApiFormAdjustFilterType,
   ApiFormFieldSet
 } from '@lib/types/Forms';
-import type { TableFieldRowProps } from '../components/forms/fields/TableField';
 import { Thumbnail } from '../components/images/Thumbnail';
 import { useGlobalSettingsState } from '../states/SettingsStates';
 import { StatusFilterOptions } from '../tables/Filter';
